requirements.txt
fastapi==0.115.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
pymongo==4.10.1
python-multipart==0.0.10